+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.6
 
+ - Drop support for 1.6
+ - Update to EffectLib 2.0 - all effect class names have changed!
  - Add "color" and "particle parameters to spells, for one-off FX color overrides.
  - Completely remove Magic.powered and Magic.protected pnodes, for good this time. (Pex!)
  - Add /wand restore command for re-creating a bound wand